requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
pytest-django==4.6.0
factory-boy==3.3.0
gunicorn==21.2.0